AJ-248: change totalBytes from int to long

+++ src/com/xxdb/data/BasicComplexMatrix.java
@@ -77,12 +77,12 @@ public class BasicComplexMatrix extends AbstractMatrix{
 	protected void readMatrixFromInputStream(int rows, int columns,	ExtendedDataInput in)  throws IOException{
 		int size = rows * columns;
 		values =new Double2[size];
-		int totalBytes = size * 16, off = 0;
+		long totalBytes = (long)size * 16, off = 0;
 		ByteOrder bo = in.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
 		while (off < totalBytes) {
-			int len = Math.min(BUF_SIZE, totalBytes - off);
+			int len = (int)Math.min(BUF_SIZE, totalBytes - off);
 			in.readFully(buf, 0, len);
-			int start = off / 16, end = len / 16;
+			int start = (int)(off / 16), end = len / 16;
 			ByteBuffer byteBuffer = ByteBuffer.wrap(buf, 0, len).order(bo);
 			for (int i = 0; i < end; i++)
 				values[i + start] = new Double2(byteBuffer.getDouble(i * 16), byteBuffer.getDouble(i*16 + 8));
