+++ src/com/xxdb/data/BasicFloatVector.java
@@ -70,13 +70,13 @@ public class BasicFloatVector extends AbstractVector{
 		int cols = in.readInt(); 
 		int size = rows * cols;
 		values = new float[size];
-		int totalBytes = size * 4, off = 0;
+		long totalBytes = (long)size * 4, off = 0;
 		ByteOrder bo = in.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
 		byte[] buf = new byte[4096];
 		while (off < totalBytes) {
-			int len = Math.min(4096, totalBytes - off);
+			int len = (int)Math.min(4096, totalBytes - off);
 			in.readFully(buf, 0, len);
-			int start = off / 4, end = len / 4;
+			int start = (int)(off / 4), end = len / 4;
 			ByteBuffer byteBuffer = ByteBuffer.wrap(buf, 0, len).order(bo);
 			for (int i = 0; i < end; i++)
 				values[i + start] = byteBuffer.getFloat(i * 4);
@@ -89,11 +89,11 @@ public class BasicFloatVector extends AbstractVector{
 	
 	@Override
 	public void deserialize(int start, int count, ExtendedDataInput in) throws IOException {
-		int totalBytes = count * 4, off = 0;
+		long totalBytes = (long)count * 4, off = 0;
 		ByteOrder bo = in.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
 		byte[] buf = new byte[4096];
 		while (off < totalBytes) {
-			int len = Math.min(4096, totalBytes - off);
+			int len = (int)Math.min(4096, totalBytes - off);
 			in.readFully(buf, 0, len);
 			int end = len / 4;
 			ByteBuffer byteBuffer = ByteBuffer.wrap(buf, 0, len).order(bo);
