+++ src/com/xxdb/data/BasicLongVector.java
@@ -69,13 +69,13 @@ public class BasicLongVector extends AbstractVector{
 		int cols = in.readInt(); 
 		int size = rows * cols;
 		values = new long[size];
-		int totalBytes = size * 8, off = 0;
+		long totalBytes = (long)size * 8, off = 0;
 		ByteOrder bo = in.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
 		byte[] buf = new byte[4096];
 		while (off < totalBytes) {
-			int len = Math.min(4096, totalBytes - off);
+			int len = (int)Math.min(4096, totalBytes - off);
 			in.readFully(buf, 0, len);
-			int start = off / 8, end = len / 8;
+			int start = (int)(off / 8), end = len / 8;
 			ByteBuffer byteBuffer = ByteBuffer.wrap(buf, 0, len).order(bo);
 			for (int i = 0; i < end; i++)
 				values[i + start] = byteBuffer.getLong(i * 8);
@@ -88,11 +88,11 @@ public class BasicLongVector extends AbstractVector{
 	
 	@Override
 	public void deserialize(int start, int count, ExtendedDataInput in) throws IOException {
-		int totalBytes = count * 8, off = 0;
+		long totalBytes = (long)count * 8, off = 0;
 		ByteOrder bo = in.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
 		byte[] buf = new byte[4096];
 		while (off < totalBytes) {
-			int len = Math.min(4096, totalBytes - off);
+			int len = (int)Math.min(4096, totalBytes - off);
 			in.readFully(buf, 0, len);
 			int end = len / 8;
 			ByteBuffer byteBuffer = ByteBuffer.wrap(buf, 0, len).order(bo);
